mern-complete

+++ src/components/WorkoutDetails.js
@@ -1,41 +1,51 @@
-import React from "react";
-import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
-import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import React from 'react';
+import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
+import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 const WorkoutDetails = ({ workout }) => {
-  const { dispatch } = useWorkoutsContext();
-  const handleClick = async () => {
-    const response = await fetch("/api/workouts/" + workout._id, {
-      method: "DELETE",
-    });
+	const { dispatch } = useWorkoutsContext();
+	const { user } = useAuthContext();
 
-    const json = await response.json();
+	const handleClick = async () => {
+		if (!user) {
+			return;
+		}
 
-    if (response.ok) {
-      dispatch({ type: "DELETE WORKOUT", payload: json });
-    }
-  };
+		const response = await fetch('/api/workouts/' + workout._id, {
+			method: 'DELETE',
+			headers: {
+				Authorization: `Bearer ${user.token}`,
+			},
+		});
 
-  return (
-    <div className="workout-details">
-      <h4>{workout.title}</h4>
-      <p>
-        <strong>Load (kg): </strong>
-        {workout.load}
-      </p>
-      <p>
-        <strong>Reps: </strong>
-        {workout.reps}
-      </p>
-      <p>
-        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
-      </p>
-      {/* delete */}
-      <span className="material-symbols-outlined" onClick={handleClick}>
-        delete
-      </span>
-    </div>
-  );
+		const json = await response.json();
+
+		if (response.ok) {
+			dispatch({ type: 'DELETE WORKOUT', payload: json });
+		}
+	};
+
+	return (
+		<div className="workout-details">
+			<h4>{workout.title}</h4>
+			<p>
+				<strong>Load (kg): </strong>
+				{workout.load}
+			</p>
+			<p>
+				<strong>Reps: </strong>
+				{workout.reps}
+			</p>
+			<p>
+				{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
+			</p>
+			{/* delete */}
+			<span className="material-symbols-outlined" onClick={handleClick}>
+				delete
+			</span>
+		</div>
+	);
 };
 
 export default WorkoutDetails;
